+++ graph-algorithms/prim/prim.c
@@ -99,7 +99,7 @@ void prim(const adj_lst_t *a,
 	  const heap_ht_t *hht,
 	  int (*cmp_wt)(const void *, const void *)){
   size_t wt_size = a->wt_size;
-  size_t vt_size = sizeof(uint64_t);
+  size_t vt_size = sizeof(size_t);
   size_t init_count = 1;
   size_t u, v;
   size_t i;
